requirements.txt
openai>=1.30.0
websockets>=12.0
pillow>=10.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
